// types/user.ts
import {Model, Document } from "mongoose"

export interface User {
    name: string;
    email: string;
    password: string;
    isAdmin?: string;
}

export interface UserDocument extends User, Document {
    matchPassword: (password: string) => Promise<Boolean>;
}

export interface UserModel extends Model<UserDocument> {}